requirements.txt
discord.py
python-dotenv
PyNaCl
ffmpeg
